refactor : Searchbar 폴더 common 폴더 하위로 이동
