+++ src/components/common/Searchbar/_variable.scss
