+++ src/components/common/Searchbar/components/CategorySelector/components/DropdownMenu/index.module.scss
@@ -1,4 +1,4 @@
-@use '@/components/Searchbar/_variable' as vars;
+@use '@/components/common/Searchbar/_variable' as vars;
 
 .dropdownMenu {
   position: absolute;
