+++ src/components/common/Searchbar/components/CategorySelector/hooks/useOpenDropdownMenu.ts
