+++ src/components/common/Searchbar/components/CategorySelector/index.tsx
