+++ src/components/common/Searchbar/components/SearchInputField/index.module.scss
@@ -1,4 +1,4 @@
-@use '@/components/Searchbar/_variable' as vars;
+@use '@/components/common/Searchbar/_variable' as vars;
 
 .input {
   width: calc(
