+++ src/components/common/Searchbar/index.module.scss
