+++ src/components/common/Toast/index.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import Image from 'next/image';
 import React, { useEffect } from 'react';
