+++ src/components/common/index.tsx
@@ -1,1 +1,2 @@
 export { default as Toast } from './Toast';
+export { default as Searchbar } from './Searchbar';
