+++ src/components/index.tsx
@@ -1,2 +1,3 @@
 export { default as Footer } from './Footer';
 export { default as Topbar } from './Topbar/index';
+export * from './common';
