+++ src/stories/Searchbar.stories.ts
@@ -1,8 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
+import { Searchbar } from '@/components';
 import { BOOK_SEARCH_CATEGORY_NAME } from '@/components/Topbar';
-
-import Searchbar from '../components/Searchbar';
 
 const meta: Meta<typeof Searchbar> = {
   component: Searchbar,
